requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
biopython=1.78=py39h3811e60_2
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py39hf3d152e_1
cycler=0.10.0=py_2
dbus=1.13.6=hfdff14a_1
expat=2.2.10=h9c3ff4c_0
fontconfig=2.13.1=hba837de_1004
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.0=h9c3ff4c_1
glib-tools=2.68.0=h9c3ff4c_1
gst-plugins-base=1.18.4=h29181c9_0
gstreamer=1.18.4=h76c114f_0
icu=68.1=h58526e2_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
kiwisolver=1.3.1=py39h1a9c180_1
krb5=1.17.2=h926e7f8_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libclang=11.1.0=default_ha53f305_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=9.3.0=hff62375_18
libgfortran5=9.3.0=hff62375_18
libglib=2.68.0=h3e27bee_1
libgomp=9.3.0=h2828fa1_18
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_openblas
libllvm11=11.1.0=hf817b99_0
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libpq=13.1=hfd2b0eb_2
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.2.0=hdc55705_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.10=h72842e0_3
lz4-c=1.9.3=h9c3ff4c_0
matplotlib=3.3.4=py39hf3d152e_0
matplotlib-base=3.3.4=py39h2fa2bec_0
mysql-common=8.0.23=ha770c72_1
mysql-libs=8.0.23=h935591d_1
ncurses=6.2=h58526e2_4
nspr=4.30=h9c3ff4c_0
nss=3.63=hb5efdd6_0
numpy=1.20.1=py39hdbf815f_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1j=h7f98852_0
pcre=8.44=he1b5a44_0
pillow=8.1.2=py39hf95b381_0
pip=21.0.1=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py39hf3d152e_7
pyqt-impl=5.12.3=py39h0fcd23e_7
pyqt5-sip=4.19.18=py39he80948d_7
pyqtchart=5.12=py39h0fcd23e_7
pyqtwebengine=5.12.1=py39h0fcd23e_7
python=3.9.2=hffdb5ce_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
qt=5.12.9=hda022c4_4
readline=8.0=he28a2e2_2
scikit-learn=0.24.1=py39h4dfa638_0
scipy=1.6.1=py39hee8e79c_0
setuptools=49.6.0=py39hf3d152e_3
six=1.15.0=pyh9f0ad1d_0
sqlite=3.35.2=h74cdb3f_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
tornado=6.1=py39h3811e60_1
tzdata=2021a=he74cb21_0
wheel=0.36.2=pyhd3deb0d_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1010
zstd=1.4.9=ha95c52a_0
